+++ test/int/test_runtime_configuration.py
@@ -10,17 +10,20 @@ def test_set_via_alter_system(cluster: Cluster):
     assert box_config["net_msg_max"] == 768
     assert box_config["checkpoint_interval"] == 3600
     assert box_config["checkpoint_count"] == 2
+    assert box_config["sql_cache_size"] == 5242880
 
     # picodata parameters names are slightly different from the tarantools
     instance.sql("ALTER SYSTEM SET iproto_net_msg_max TO 100 FOR ALL TIERS")
     instance.sql("ALTER SYSTEM SET memtx_checkpoint_interval TO 100 FOR ALL TIERS")
     instance.sql("ALTER SYSTEM SET memtx_checkpoint_count TO 100 FOR ALL TIERS")
+    instance.sql("ALTER SYSTEM SET sql_storage_cache_size_max TO 100 FOR ALL TIERS")
 
     # parameters values changed
     box_config = instance.eval("return box.cfg")
     assert box_config["net_msg_max"] == 100
     assert box_config["checkpoint_interval"] == 100
     assert box_config["checkpoint_count"] == 100
+    assert box_config["sql_cache_size"] == 100
 
     # box settings isn't persistent, so it should be reapplied
     instance.restart()
@@ -31,6 +34,7 @@ def test_set_via_alter_system(cluster: Cluster):
     assert box_config["net_msg_max"] == 100
     assert box_config["checkpoint_interval"] == 100
     assert box_config["checkpoint_count"] == 100
+    assert box_config["sql_cache_size"] == 100
 
     # bad values for parameters shouldn't pass validation
     # stage before creating DML from ir node
@@ -203,3 +207,56 @@ cluster:
     assert blue_config["checkpoint_interval"] == 3600
     red_config = red_instance.eval("return box.cfg")
     assert red_config["checkpoint_interval"] == 3600
+
+
+def test_cache_capacity(cluster: Cluster):
+    i1 = cluster.add_instance()
+
+    i1.sql("ALTER SYSTEM SET sql_storage_cache_count_max = 1")
+
+    cache_info = i1.eval("return box.info.sql()")
+    assert cache_info["cache"]["stmt_count"] == 0
+
+    # random sql that inserts to tarantool cache
+    i1.sql("SELECT * FROM _pico_instance")
+
+    cache_info = i1.eval("return box.info.sql()")
+    assert cache_info["cache"]["stmt_count"] == 1
+
+    i1.sql("SELECT * FROM _pico_replicaset")
+
+    cache_info = i1.eval("return box.info.sql()")
+    assert cache_info["cache"]["stmt_count"] == 1
+
+    i1.sql("ALTER SYSTEM SET sql_storage_cache_count_max = 2")
+
+    i1.sql("SELECT * FROM _pico_replicaset")
+
+    # select from replicaset already cached
+    cache_info = i1.eval("return box.info.sql()")
+    assert cache_info["cache"]["stmt_count"] == 1
+
+    i1.sql("ALTER SYSTEM SET sql_storage_cache_count_max = 3")
+
+    i1.sql("SELECT * FROM _pico_instance")
+
+    cache_info = i1.eval("return box.info.sql()")
+    assert cache_info["cache"]["stmt_count"] == 2
+
+    i1.sql("SELECT * FROM _pico_tier")
+
+    cache_info = i1.eval("return box.info.sql()")
+    assert cache_info["cache"]["stmt_count"] == 3
+
+    # cache can shrink
+    i1.sql("ALTER SYSTEM SET sql_storage_cache_count_max = 1")
+
+    cache_info = i1.eval("return box.info.sql()")
+    assert cache_info["cache"]["size"] == 2442
+
+    i1.sql("SELECT * FROM _pico_tier")
+
+    # if size doesn't changed, then query was in cache, and it's true,
+    # because of LRU
+    cache_info = i1.eval("return box.info.sql()")
+    assert cache_info["cache"]["size"] == 2442